Refactored GeneSearchService to move Solr querying functionality to the Dao class.
